Finished mysql integ proj

+++ 03-mysql-bookstore/pkg/models/book.go
@@ -32,13 +32,13 @@ func GetAllBooks() []Book {
 	return books
 }
 
-func GetBookById(Id int) (*Book, *gorm.DB) {
+func GetBookById(Id int64) (*Book, *gorm.DB) {
 	var getBook Book
 	db := db.Where("ID=?", Id).Find(&getBook)
 	return &getBook, db
 }
 
-func DeleteBookById(Id int) Book {
+func DeleteBookById(Id int64) Book {
 	var book Book
 	db.Where("ID=?", Id).Delete(book)
 	return book
